Add an abstraction layer to influx command
This adds an abstraction layer to the influx command, in preparation for
adding an MQTT target. All ANT+ handling is moved into
`base/datatarget.py` and only handling of Influx specifics is left in
the influx command itself.

The logic and code is essentially unchanged, although it has been moved
around some, and changed into a class structure. The only visible change
should be a slight formatting change in the influx command help,
grouping the influx specific options.
